refactor: consolidate Softplus into ActivationConversion and ActivationLowering
Move Softplus conversion and lowering patterns from standalone files into
the shared Activation files for consistency with Sigmoid/Silu pattern.

Changes:
- OnnxToHip: Move SoftplusToHip pattern into ActivationConversion.cpp
- HipToLLVM: Move SoftplusOpLowering into ActivationLowering.cpp
- Delete standalone SoftplusConversion.cpp and SoftplusLowering.cpp
- Update CMakeLists.txt to remove deleted files
- Remove forward declarations from OnnxToHipUtils.h and HipToLLVMUtils.h
- Remove populate function calls from OnnxToHip.cpp and HipToLLVM.cpp
- Update tests to check for wrap_miopenActivationForward instead of wrap_softplus
- Update tests to verify activation_mode=3 (HIPDNN_EP_ACTIVATION_SOFTPLUS)

All 81 lit tests pass (100%).
