--- lib/Conversion/HipToLLVM/SoftplusLowering.cpp
@@ -1,85 +1,0 @@
-/*
- * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
- * Licensed under the MIT License.
- */
-
-#include "HipToLLVMUtils.h"
-
-namespace mlir {
-namespace hip {
-namespace {
-
-// hip.softplus(ctx, input, output)
-//   -> wrap_softplus(state, input, output, num_elements, data_type)
-// Supports both static and dynamic shapes (computes num_elements at runtime).
-struct SoftplusOpLowering : public ConvertOpToLLVMPattern<SoftplusOp> {
-  using ConvertOpToLLVMPattern::ConvertOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(SoftplusOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    Location loc = op.getLoc();
-    ModuleOp module = op->getParentOfType<ModuleOp>();
-    Type ptrType = getPtrType();
-    Type i32Type = rewriter.getI32Type();
-    Type i64Type = rewriter.getI64Type();
-
-    // Helper to create i64 constants
-    auto createI64Const = [&](int64_t value) -> Value {
-      return LLVM::ConstantOp::create(rewriter, loc, i64Type,
-                                      rewriter.getI64IntegerAttr(value));
-    };
-
-    // Extract pointers using alignedPtr (respects memref.view offsets)
-    Value statePtr = adaptor.getCtx();
-    Value inputPtr = extractMemRefPtr(adaptor.getInput(), rewriter, loc);
-    Value outputPtr = extractMemRefPtr(adaptor.getOutput(), rewriter, loc);
-
-    auto outputType = cast<MemRefType>(op.getOutput().getType());
-
-    // Compute num_elements (supports dynamic shapes)
-    Value numElements = createI64Const(1);
-    MemRefDescriptor outputDesc(adaptor.getOutput());
-
-    for (auto dimIdx : llvm::seq<int64_t>(outputType.getRank())) {
-      Value dimSize;
-      if (outputType.isDynamicDim(dimIdx)) {
-        dimSize = outputDesc.size(rewriter, loc, dimIdx);
-      } else {
-        dimSize = createI64Const(outputType.getDimSize(dimIdx));
-      }
-      numElements = LLVM::MulOp::create(rewriter, loc, numElements, dimSize);
-    }
-
-    // Get data type enum
-    int64_t dataType = getHipdnnDataType(outputType.getElementType());
-    Value dataTypeVal = createI64Const(dataType);
-
-    // int wrap_softplus(RuntimeState* state, void* input, void* output,
-    //                   int64_t num_elements, int64_t data_type)
-    SmallVector<Type, 5> paramTypes = {ptrType, ptrType, ptrType, i64Type,
-                                       i64Type};
-
-    FailureOr<LLVM::LLVMFuncOp> funcOp = LLVM::lookupOrCreateFn(
-        rewriter, module, kWrapSoftplus, paramTypes, i32Type);
-    if (failed(funcOp))
-      return failure();
-
-    SmallVector<Value, 5> args = {statePtr, inputPtr, outputPtr, numElements,
-                                  dataTypeVal};
-
-    LLVM::CallOp::create(rewriter, loc, *funcOp, args);
-    rewriter.eraseOp(op);
-    return success();
-  }
-};
-
-} // namespace
-
-void mlir::hip::populateSoftplusLoweringPatterns(
-    const LLVMTypeConverter &converter, RewritePatternSet &patterns) {
-  patterns.add<SoftplusOpLowering>(converter);
-}
-
-} // namespace hip
-} // namespace mlir
