+++ lib/Conversion/OnnxToHip/ActivationConversion.cpp
@@ -67,11 +67,40 @@ SigmoidToHip::matchAndRewrite(mlir::Operation *op,
   return mlir::success();
 }
 
+/// onnx.Softplus -> hip.softplus
+struct SoftplusToHip : public mlir::RewritePattern {
+  SoftplusToHip(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Softplus", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override;
+};
+
+mlir::LogicalResult
+SoftplusToHip::matchAndRewrite(mlir::Operation *op,
+                               mlir::PatternRewriter &rewriter) const {
+  auto ctxOrFailure = getContextArg(op, rewriter);
+  if (mlir::failed(ctxOrFailure))
+    return mlir::failure();
+  mlir::Value context = *ctxOrFailure;
+
+  mlir::Location loc = op->getLoc();
+  mlir::Value input = op->getOperand(0);
+  auto resultType =
+      mlir::cast<mlir::RankedTensorType>(op->getResult(0).getType());
+  mlir::Value init = createEmptyTensor(rewriter, loc, resultType, input);
+  auto hipOp = mlir::hip::SoftplusOp::create(rewriter, loc, resultType, context,
+                                             input, init);
+  rewriter.replaceOp(op, hipOp->getResult(0));
+  return mlir::success();
+}
+
 } // namespace
 
 void mlir::hip::populateActivationConversionPatterns(
     RewritePatternSet &patterns, MLIRContext *ctx) {
-  patterns.add<SoftmaxToHip, SigmoidToHip>(ctx);
+  patterns.add<SoftmaxToHip, SigmoidToHip, SoftplusToHip>(ctx);
 }
 
 } // namespace hip
