+++ lib/Conversion/OnnxToHip/OnnxToHipUtils.h
@@ -139,8 +139,6 @@ void populateGatherConversionPatterns(RewritePatternSet &patterns,
                                       MLIRContext *ctx);
 void populateReshapeConversionPatterns(RewritePatternSet &patterns,
                                        MLIRContext *ctx);
-void populateSoftplusConversionPatterns(RewritePatternSet &patterns,
-                                        MLIRContext *ctx);
 
 } // namespace hip
 } // namespace mlir
