+++ test/lit/e2e/test_softplus_model.mlir
@@ -10,13 +10,13 @@
 // 4. convert-hip-to-llvm: HIP ops → LLVM runtime calls
 // 5. generate-interface: Create inference_init/compute/cleanup/metadata
 //
-// Softplus uses wrap_softplus() which delegates to MIOpen's
-// wrap_miopenActivationForward() with HIPDNN_EP_ACTIVATION_SOFTPLUS.
+// Softplus directly uses MIOpen's wrap_miopenActivationForward()
+// with HIPDNN_EP_ACTIVATION_SOFTPLUS (activation_mode=3).
 
 // CHECK: module attributes {
 // CHECK-SAME: hipdnn.input_count = 1
 // CHECK-SAME: hipdnn.output_count = 1
-// CHECK: llvm.func @wrap_softplus
+// CHECK: llvm.func @wrap_miopenActivationForward
 // CHECK: llvm.func @inference_init
 // CHECK: llvm.func @inference_compute
 // CHECK: llvm.func @inference_cleanup
